Clone without diffing
Selects on the target side can be increasingly expensive if we don't have an index by shard or keyspace id. In the case it will be much faster to clone without diffing. This won't handle deleted rows on the source so can only be used if we do consistent clones with nothing on the target side.