We should use READ_COMMITTED for most things
READ_COMMITTED should be fine because while the default (REPEATABLE_READ) protects against a few read anomalies we take great care in cloner to make sure there are no concurrent writes to the same data. the only reads we do out of the target database is when we diff a chunk before we repair it but that should never be concurrent with any other writes to the same chunk.